Add Dynatrace REST API client
This adds a client interface for the Dynatrace REST API that provides
the following two queries: GetVersionForLatest, GetVersionForIp.

There is a proper implementation for GetVersionForIp which uses the
existing API endpoint for listing monitored hosts.
GetVersionForLatest is not implemented yet, it returns a constant
version string, because the API is not yet implemented on the server
side.