Properly watch for OneAgent pods recreation
The client.List() function falls back to a cache implementation after the initial attempt.

However, the client's cached implementation fails to handle complex queries, so rather than the function falling back to use the normal client, it just complains that the query is not supported.

On this commit, we query all the pods, and filter them ourselves as a workaround.